requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
numpy
tiktoken
google-generativeai
chromadb
uuid
discord
python-dotenv